WIP: messy draft of removing getPhoneHash
