+++ packages/sdk/base/package.json
@@ -26,7 +26,6 @@
     "@celo/flake-tracker": "0.0.1-dev",
     "@celo/typescript": "0.0.1",
     "bignumber.js": "^9.0.0",
-    "elliptic": "^6.5.4",
-    "web3-utils": "1.3.6"
+    "elliptic": "^6.5.4"
   }
 }
